chore: remove import + move nav into a dir
