+++ apps/web/src/app/layout.tsx
@@ -1,11 +1,11 @@
-import { OG_URL, tagline } from './metadata';
-import { getStaticParams } from '~/locales/server';
-import { Inter } from 'next/font/google';
-import { Providers } from './[locale]/providers';
-import { Navigation } from '~/components/navigation';
 import { Toaster } from '@repo/ui/components/toaster';
 import { Analytics } from '@vercel/analytics/react';
+import { Inter } from 'next/font/google';
+import { Navigation } from '~/components/Navigation';
+import { getStaticParams } from '~/locales/server';
 import '../styles/globals.css';
+import { Providers } from './[locale]/providers';
+import { OG_URL, tagline } from './metadata';
 
 export function generateStaticParams() {
   return getStaticParams();
